Updating Signals
Finished switching all the .Connect() calls over to the Event Handler syntax and updating the handler methods to account for Godot 4's C# Signal API standardizing on 64-bit data types.

+++ scripts/nodes/cod/ctl/meta/ContractTypeEntry.cs
@@ -7,10 +7,10 @@ namespace OCSM.Nodes.CoD.CtL.Meta
 {
 	public partial class ContractTypeEntry : BasicMetadataEntry
 	{
-		protected override void entrySelected(int index)
+		protected override void entrySelected(long index)
 		{
 			var optionsButton = GetNode<OptionButton>(NodePathBuilder.SceneUnique(ExistingEntryName));
-			var name = optionsButton.GetItemText(index);
+			var name = optionsButton.GetItemText((int)index);
 			if(metadataManager.Container is CoDChangelingContainer ccc)
 			{
 				if(ccc.ContractTypes.Find(ct => ct.Name.Equals(name)) is ContractType contractType)
