+++ scripts/nodes/dnd/fifth/meta/ClassEntry.cs
@@ -6,10 +6,10 @@ namespace OCSM.Nodes.DnD.Fifth.Meta
 {
 	public partial class ClassEntry : FeaturefulMetadataEntry
 	{
-		protected override void entrySelected(int index)
+		protected override void entrySelected(long index)
 		{
 			var optionsButton = GetNode<ClassOptionsButton>(NodePathBuilder.SceneUnique(ExistingEntryName));
-			var name = optionsButton.GetItemText(index);
+			var name = optionsButton.GetItemText((int)index);
 			if(metadataManager.Container is DnDFifthContainer dfc)
 			{
 				if(dfc.Classes.Find(c => c.Name.Equals(name)) is Class clazz)
