+++ scripts/nodes/dnd/fifth/meta/RaceEntry.cs
@@ -6,10 +6,10 @@ namespace OCSM.Nodes.DnD.Fifth.Meta
 {
 	public partial class RaceEntry : FeaturefulMetadataEntry
 	{
-		protected override void entrySelected(int index)
+		protected override void entrySelected(long index)
 		{
 			var optionsButton = GetNode<RaceOptionsButton>(NodePathBuilder.SceneUnique(ExistingEntryName));
-			var name = optionsButton.GetItemText(index);
+			var name = optionsButton.GetItemText((int)index);
 			if(metadataManager.Container is DnDFifthContainer dfc)
 			{
 				if(dfc.Races.Find(r => r.Name.Equals(name)) is Race race)
